Link Refatorado JS
O link para sugestão de email que era fixo no html passou a ser inserido via JS para evitar que fique vazio.
Repair: Add text to the a element or the title attribute of the a element or, if an image is used within the anchor, add Alt text to the image.